Always close backend device before program exit
If there's a host side crash or exception there is some device cleanup /
shutdown that doesn't automatically happen and will leave the device in
a hung state for the next run.

This change registers an atexit handler so that in most situations a
host side crash will not result in a hung device.