Workaround multi-device system descs by only observing the first device
Until we have system desc slicing #113, by default the compiler will
only construct a `DeviceAttr` that observes the first chip on the system
to enable running on n300.

Once we have system desc slicing, we can change it back, and the user
will have to explicitly generate a system desc that partitions out a
single device.