Create tensor grids with same rank as device grid bug fix #500
Previously we were just creating tensor layout grids with rank 1 or 2
depending on the tensor's rank.  This however is incorrect, the tensor
grid must be of the same rank as the device grid.

The fix is to use the device grid's rank in the layout type converter to
ensure that by default a tensor layout gets a grid of equivalent rank.