Convert to `row_major` layout before calling `reshape`
TTNN supports two kinds of reshapes - row and tile. row reshape works
as expected as long as we don't change last dim.
On the other hand tile reshape requires that output `W` and `H` dims are tile aligned.

For example this `reshape` in tile layout will fail:
1*12*32 -> 12*32

Error that we get is `Unable to reshape a tensor in TILE_LAYOUT
to non-tile height and width! Please convert the tensor to
ROW_MAJOR_LAYOUT first.`

In addition when we managed to do a `reshape` on tile tensor we
found that output tensor did not look the way we expected it to look.
For example:

If we have input row major tensor 4x4 which contains numbers from 0 .. 15
and we reshape it to 2x8 we will get [[0,1 .. 6,7], [8,9 .. 14,15]].
The output is correct.

If we take the same input tensor and convert it to tile layout
and call reshape to Shape((2, 8), (32, 32)) and convert output tensor to
row_major we will get tensor which looks like this:
[[0,1,2,3,0,0,0,0],[4,5,6,7,0,0,0,0]].This is different then row major
reshape.

In order to unblock forge for `reshape` op we will convert all input
tensors to row before calling reshape, and then we will convert output
tensor back to original layout.

Once we have understanding how reshape works for tile we will remove
this and implement proper fix.