Introduce FillOp + lowering ttir.constant to ttmetal (#771)
Introduce FillOp + lowering ttir.constant to ttmetal

- Introduce `ttir.fill` op
- Introduce pass the converts `ttir.constant` into an empty + fill
- Support lowering `ttir.fill` to `ttmetal.host_write`

The purpose of this change is to support lowering of `ttir.constant`
into the ttmetal backend.  An issue that arises with `ttir.constant`
is that the allocator pass only knows how to convert `ttir.empty`'s into
`ttir.allocate`, and we probably don't want to support lowering
`ttir.constant` in the same way because otherwise we'd have to store the
constant value attribute on the `ttir.allocate` op which would be an
awkward place to put it.  The best solution I could come up with is to
create a `ttir.fill` op, which is essentially just a DPS version of
`ttir.constant`, but now more naturally fits the allocation pass.
Further, this also better fits lowering to ttmetal host_write op
because it too expects a pre-allocated device destination, i.e. DPS
style.

Closes #758