Docker build as part of workflow (#928)
Updating the workflow to make the Docker image build the first step in the build and test pipeline. Instead of triggering Docker builds manually, the workflow will check if the required Docker images already exist on GHCR (GitHub Container Registry). If they don’t, the workflow will initiate the build.

How do we determine which Docker image is needed?

The image tag is based on a hash of the Dockerfiles and the environment CMake file. If any of these change, a new Docker image build will be triggered. Although we could refine this to only trigger builds for specific changes (e.g., LLVM version in the environment file), for now, any change will trigger a new build. See get-docker-tag.sh

If a Docker image with the calculated tag already exists, the job will quickly complete and pass the image tag to the subsequent build and test jobs, which will use this specific image instead of the "latest" tag. This ensures the workflow always uses the correct Docker image.

How do we know which image to use in IRD?

Each time a Docker image is built, it will also be tagged as "latest." In most cases, this "latest" image will work for IRD. However, if it doesn’t, developers can run the get-docker-tag.sh script to retrieve the exact image tag needed for their code and use that in IRD.