[Optimizer] L1 Interleaved policy that solves simple fork-joins (#1501)
This PR introduces new MemoryLayoutAnalysis policy as an alternative to
the GreedyL1Interleaved policy with the goal of solving simple
fork-joins. Fork-join is considered to be simple if there is no need for
DRAM spill in its execution. In this policy, we want to make sure that
we always solve simple fork-joins. Furthermore, if DRAM spill is
necessary, this policy will not produce globally optimal solution.