Add l1 small size as an optional parameter to open device (#1781)
L1 small size is a device property and needs to be set when opening
device, adding a parameter so that the user can toggle this as needed.

A future improvement would be to expose an API to the user to calculate
the l1 small size directly, or maybe have a field in the device attr in
the mlir that the compiler calculates and populates.