Fixing result type missmatch when using typecast
In certain cenarios we can end up with elementtype missmatch between
return op and func op return types. This happens because of how we
convert builtin mlir data types to tt data types.

There is an issue created for this a while ago which should add
conversion pass for element types.
https://github.com/tenstorrent/tt-mlir/issues/1277