Infer shape in builder from golden output
This change streamlines our bindings within the TTIR builder by having
the output shape passed to the TTIR op be computed from the golden
(torch) output. This both makes adding new ops easier, and reduces the
chance of a bug in how we manually calculate shapes.

During testing, I also discovered we have no single test for a matmul
through the builder, so I added one to show that the new feature works