Adding ttnn.to_dtype op in TTNN dialect (#2055)
In this pull request (PR), I am introducing the `ttnn.to_dtype`
operation for data type casting on the host within the TTNN dialect.

The existing `ttnn.typecast` function only supports data type casting on
a device, which has led to failures when casting on the host. To address
this, I have also restructured the `TTNNDecomposeLayouts.cpp` file to
accommodate these changes. Additionally, I have included compiler and
silicon tests for scenarios where the inputs are located on the host.

Since we are now managing this behavior within the compiler, there is no
longer a need for runtime workarounds. As a result, I have removed those
workarounds in this PR.

This improvement will benefit `emitC`, which previously lacked a runtime
workaround and caused several tests to fail. Now, with these changes, we
can ensure better compatibility and functionality.

I created tests for Silicon from the transform pass. Please review it,
and if anyone has a better method for running this on Silicon, your
input would be greatly appreciated. Comments are welcome! :)

With this change, we are fixing a couple of bugs noticed:
Closes #1938
Closes #1733
Closes #1658