Interface for DPS op creation (#1921)
The goal of this PR is to simplify the creation of DPS ops. Most DPS ops
follow the same pattern, where op is defined in TableGen as:
```
def DPSOp {
    ...
    let arguments = (ins
                                       AnyRankedTensor:$input1,
                                       ...,
                                       AnyRankedTensor:$inputN,
                                       AnyRankedTensor:$output,
                                       Attribute:$attr1,
                                       ...,
                                       Attribute:$attrM);
    ...
}
```
with the appropriate builder where arguments follow the order of
operands and attributes in the op definition. Frontends from which we
lower to TTIR are non-DPS, so when we lower from those frontends to TTIR
we end up (in the best case) with the following code:

```cpp
    RankedTensorType outputType = mlir::cast<RankedTensorType>(
        this->getTypeConverter()->convertType(srcOp.getResult().getType()));
    tensor::EmptyOp output = rewriter.create<tensor::EmptyOp>(
        adaptor.getLoc(), outputType.getShape(), outputType.getElementType());
    rewriter.replaceOpWithNewOp<DestOp>(
        srcOp, outputType, adaptor.getInput1(),..., adaptor.getInputN(), output, adaptor.getAttr1(),..., adaptor.getAttrM());
```

I mentioned "in the best case" specifically because this whole process
is repeating and error-prone, so even though it's not a big block of
code, there are still a lot of places where it could get wrong. There
are a lot of places where it did get wrong, a most notable example is
the double type conversion of output type that happened during the
StableHLOToTTIR conversion, and that's actually the code that has
compiled and executed as expected because the way type converter is
currently(!) defined doesn't involve any transitive conversion, but
whoever worked with DPS ops in more complicated decompositions knows how
easy is to make a mistake that takes some time to find.

Proposed functions `createDPSOp` and `replaceOpWithNewDPSOp` should be
almost a drop-in replacement for `createOp` and `replaceOpWithNewOp`
respectively. So for the aforementioned example, the replacement would
look like this.

```cpp
    RankedTensorType outputType = mlir::cast<RankedTensorType>(
        this->getTypeConverter()->convertType(srcOp.getResult().getType()));
   replaceOpWithNewDPSOp<DestOp>(
        rewriter, srcOp, outputType, adaptor.getInput1(),..., adaptor.getInputN(), adaptor.getAttr1(),..., adaptor.getAttrM());
```

These functions are template functions, that rely on the previously
mentioned structure (all inputs, output, all attributes). As it's a
template function, it can be challenging to recognize where operands end
(or attributes begin), because we can't use RTTI. So the way I currently
define an operand is like this:

```cpp
(std::is_convertible_v<T, mlir::Value> ||
 std::is_convertible_v<T, mlir::ValueRange>) &&
!std::is_convertible_v<T, mlir::TypedValue<mlir::tt::DeviceType>>
```

It's not ideal, but this is something that works for most of our use
cases (note that `device` is the only exception, because it's
convertible to `Value` but it behaves more like an attribute in DPS ops
in which it's defined). This definition doesn't aim to be the most
comprehensive, just to cover most cases. At the end of the day these
functions are here for convenience (with all the benefits that come with
DRY), we can still fall back to the old way of defining when we
encounter some cases that aren't covered with this pattern (there are
some cases not covered still left, most notably DPS ops with multiple
results). I'm open to suggestions for improvement, especially for the
way how to find the boundary between operands and attributes.

Closes https://github.com/tenstorrent/tt-mlir/issues/1671