TTRT incremental build fix (#2188)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/2187

### Problem description
The incremental build of TTRT is not working properly. As I understand
the root cause is that `python -m pip wheel` will run only when one of
the files in `runtime/tools/python` is touched, but it should also run
in case `TTRuntime` or some of its dependencies have changed because we
need to rebuild `ttrt.runtime._C` and `ttrt.binary._C` shared libraries.

### What's changed
Added new dependencies. This is a somewhat hacky solution, so I'm very
open to suggestions on how to improve this, or even take a different
approach.

### Checklist
- [ ] New/Existing tests provide coverage for changes