Static build (#2108)
Recent refactoring of ttmlir cmake files has broken the macOS build.
During the support fix for macOS, there are also a number of other
cleanups that were made:
- Switch to static build wholesale, this *should* resolve a number of
weird runtime errors of unregistered dialects/passes or already
registered dialects/passes
- Remove `lib/SharedLib` subdir. Now that we only build shared libs,
this seemed redundant so this cmake file is collapsed into the dir above
it. This also fixed a link path error that `ttmlir-opt` runs into at
runtime because it looks in `build/lib` and doesn't expect it in
`build/lib/SharedLib`.
- Following the above bullet, we no longer need to use `empty.cpp`, we
can just use `RegisterAll.cpp` as the base source file.
- Use `-Wl,-all_load` on macOS instead of `-Wl,--whole-archive`
- Move `TTMLIRRuntime` target to `runtime/lib` and merge it with
`TTRuntime`, this also simplifies the link step of having to re-export
the API.
- Only `ExternalProject_Add(tt-metal)` if runtime is enabled