Exposed the createOwnedTensor API in the runtime to the frontends (#2204)
As outlined in issue https://github.com/tenstorrent/tt-mlir/issues/2203,
the fact that there is no way to create an owned tensor has posed a
significant problem to the tt-xla frontend and the multichip effort.
This PR exposes that functionality, allowing creating of tensors in
which the ttnn runtime owns the tensor data.

Fixes https://github.com/tenstorrent/tt-mlir/issues/2203