ConstantOp in TTNN and runtime (#2075)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/1584

### Problem description
We didn't have a way to create an arbitrary constant tensor. This PR
introduces constant 'op' in the TTNN dialect and runtime.

### What's changed
Here is a good description of the approach
https://github.com/tenstorrent/tt-mlir/issues/1584#issuecomment-2543314700.
To summarize, there isn't a constant op in `TTNN` lib, but the way to
deal with it is to model it as a `ConstantOp` in TTNN dialect and use
the provided `ttnn::Tensor` constructor for creation of tensor in the
runtime.

Previously, there was a `ttir.constant` that would lower to `ttnn.full`
if provided value is a single number that broadcasts to full dimension.
This PR doesn't change that case, and still tries to lower to
`ttnn.full` whenever possible, but otherwise lowers to `ttnn.constant`.