Add argument type and pass which can populate them (#2049)
Issue: https://github.com/tenstorrent/tt-mlir/issues/2032

- Create `TTPopulateArgumentTypes` pass which can be used to add
argument types to the functions of a module.
- Provide `TTIRToTTNNBackedPipeline` option which can be used to
populate the argument types of all public functions in the MLIR module.
`TTPopulateArgumentTypes` is added to `TTIRToTTNNBackendPipeline`
    - Argument types are stored in the `FuncOp`s `arg_attrs` attribute. 
- If this option is unpopulated it defaults to an empty
`llvm::StringMap`, in which case the pass that populates the attributes
returns immediately - doing nothing.
- If any arg attributes are populated before running this pass, they
will remain. **Except** any existing `tt.argument_type` attributes. If
this pass is run, the existing `tt.argument_type` attributes will be
overwritten by the ones specified in `TTIRToTTNNBackendPipelineOptions`.
- This means that a user of this pipeline is free to add these
attributes on their own if they wish. To avoid overwriting these values
they can simply not set the argument map in
`TTIRToTTNNBackendPipelineOptions` as the default (empty map) causes the
pass to not modify the module.