Decouple ttmetal and ttnn flatbuffer types (#2217)
### Problem description
We have been using a shared types.fbs file as an interface between mlir
and runtime types. This caused ttmetal types and ttnn types to be
coupled and it makes it hard to model backend-specific constructs. We
also needed to add workarounds (e.g. TensorMemoryLayout::NONE) for
compatibility between both backends. As we start expanding work on
direct to metal as well as multichip work on ttnn, it would be desirable
to separate backend-specific types.

### What's changed
This PR is a refactors the mlir to flatbuffer interface as well as
runtime. Specific changes include:

* Moved backend-specific types out of `Common/types.fbs` into
`ttmetal/types.fbs` and `ttnn/types.fbs`.
* Refactored ttnn `TensorDesc` to contain `StorageType` and
`MemoryConfig`. Also moved core range into `ShardSpec` to match the
modelling in ttnn.
* Removed `TensorMemoryLayout`, `DistributionStrategy` from ttmetal
TensorDesc.
* Removed `TensorMemoryLayout::NONE` from flatbuffer entirely.
* Canonicalized `createMemoryConfig` API in ttnn backend, now the API
just needs a `::tt::target::ttnn::MemoryConfig` parameter.
* Updated runtime code to follow established coding guidelines,
specifically changing `not` to `!`, `and` to `&&` etc.
* Update ttnn backend to use ttnn type aliases when possible, instead of
a mixture of say `::ttnn::MemoryConfig` and
`::tt::tt_metal::MemoryConfig`.

### Future Improvements
* This PR supports 1-to-1 mapping between flatbuffer and runtime. It
would also be desirable to have a 1-to-1 mapping between compiler and
flatbuffer (FYI @odjuricicTT). Currently many types/attributes are still
coupled on the compiler end, and it would be nice to separate them and
start adding backend-specific attributes to support 1-to-1 mapping and
consistency throughout the stack across different backends.
  
* Currently tensor `storageType` is hardcoded to either owned (for host
tensors) or device (for device tensors). Ideally we should have compiler
support for this to dynamically program the storage type. We could add a
storage type to the layout attributes in ttnn dialect such that when a
tensor is an output of say a mesh_shard op, it'll change its storage to
multi_device_host or back to owned (if it's a shard to full).