improving ttkernel conversion tests: part 1 (#2282)
### Ticket
This is part 1 of #2068. More dialect ops will be covered but I would
like to get the new lit test form into main as an example for future
tests.

### Problem description
Existing lit tests for ttkernel seemed to be in a form of one func that
tests several ops in one monolithic block.

### What's changed

- the test has been split into func-per-op, in the style of MLIR's own
dialects like arith
- better regex matching is used to validate the ordering of param
passing in lowered form
- several ops confirmed to be unused/never tested removed from the
dialect