Create `ttnn.neg` as non-DPS op (#2284)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/2283

### Problem description
In one place in TTIR->TTNN conversion, `NegOp` is created as DPS op,
even though it isn't.

### What's changed
Since `NegOp` is elementwise op and they were recently changed from DPS
to non DPS op it was an error to create it like a DPS op, because it
took empty tensor as an input. Added `static_assert`s in DPSOp creation
functions to check if target op is actually a DPS op.

### Checklist
- [ ] New/Existing tests provide coverage for changes