Removing simple builders from eltwise binary and ternary ops. (#2295)
### Ticket
Closes https://github.com/tenstorrent/tt-mlir/issues/2294

### Problem description
Frontends reported a few issues when the broadcast is applied to eltwise
ops.

### What's changed
We removed the simple builders for eltwise binary and ternary ops. This
is the main reason for bugs in the code, hence we are temporarily
removing them to unblock the uplift. We have created the following issue
to follow up on these builders later:
https://github.com/tenstorrent/tt-mlir/issues/2263

### Checklist
- [x] New/Existing tests provide coverage for changes