Check and update return sharding (#2298)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/2297

### Problem description
PJRT expects the return values to be sharded instead of being merged
into a tensor with a single buffer if there is sharding attribute in
return value.

### What's changed
This PR provides special care for the return values just like we do for
input tensors. It checks the return value attribute and if the sharding
attribute is there, mark the returning mesh_shard op as manual such that
runtime will not perform any concat operation and returns multiple
buffers.

### Checklist
- [V] New/Existing tests provide coverage for changes