Fixing TTNN NOperand trait verify method visibility. (#2334)
### Ticket
Closes https://github.com/tenstorrent/tt-mlir/issues/2333

### Problem description
TTNN trait NOperandTTNN fails to call verifyTrait method:

https://github.com/tenstorrent/tt-mlir/blob/7bc75fc277ac11c6c0b2cfaeba25ad3d42af04be/include/ttmlir/Dialect/TTNN/IR/TTNNTraits.h#L76C7-L76C19

The issue is the visibility of the verifyTrait method, in the current
code, it is private, hence it can't be called by the framework.

### What's changed
Changed the visibility of verifyTrait from private to public, so the
verifyTrait method is executed. Added tests to verify this behaviour.

### Checklist
- [x] New/Existing tests provide coverage for changes