Try to optimize docker build image (#2293)
Found redundant step in CI docker image.

Previously we run application build as part of docker image build, to
test the docker image and dependancies.
We can now remove this step as docker now builds as part of build and
test CI and we will run app build and test with new image in following
steps.