DenseResourceElementsAttr as constant value E2E (#2381)
### Ticket
Closes https://github.com/tenstorrent/tt-mlir/issues/2378

### Problem description
When the value of a `stalbehlo/ttir/ttnn` constant op is a
`DenseResourceElementsAttr` instead of a `DenseElementsAttr`, we fail to
lower stablehlo to ttir. We also do not yet handle the case where the
value of a constant is in a `DenseResourceElementsAttr` when generating
the flatbuffer.

### What's changed
- Legalize lowering of `stablehlo.constant` when the value type is
`DenseResourceElementsAttr`
- Handle case where value type is `DenseResourceElementsAttr` when
generating flatbuffer
- Add verifier to `ttnn.constant` to ensure that the `ElementsAttr`
holding the constant data is either `DenseResourceElementsAttr` or
`DenseElementsAttr`