1D Convolution where input spatial dimension isn't 2 (#2358)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/2357

### Problem description
In TTIR to TTIR decomposition, `Legalize1DConvolutionPattern` is not
currently supporting spatial dimensions other than 2.

### What's changed
This case is already supported and the check isn't necessary. After
reshaping input and output to match 2D `ttir.convolution` op and the new
`ttir.convolution` op is picked up by the `ConvolutionToConv2dPattern`,
the dimensions are permuted and handled correctly.
Added a test for this case.

### Checklist
- [x] New/Existing tests provide coverage for changes